cosmetic

+++ src/openMVG/sfm/sfm_data_io_ply.hpp
@@ -65,7 +65,7 @@ inline bool Save_PLY(
           if (sfm_data.IsPoseAndIntrinsicDefined(view.second.get()))
           {
             const geometry::Pose3 pose = sfm_data.GetPoseOrDie(view.second.get());
-            stream << pose.center().transpose()
+            stream << pose.center().transpose().cast<float>()
               << " 0 255 0" << '\n';
           }
         }
@@ -78,7 +78,7 @@ inline bool Save_PLY(
         for (Landmarks::const_iterator iterLandmarks = landmarks.begin();
           iterLandmarks != landmarks.end();
           ++iterLandmarks)  {
-          stream << iterLandmarks->second.X.transpose() << " 255 255 255" << '\n';
+          stream << iterLandmarks->second.X.transpose().cast<float>() << " 255 255 255" << '\n';
         }
       }
 
@@ -89,7 +89,7 @@ inline bool Save_PLY(
         for (Landmarks::const_iterator iterGCP = landmarks.begin();
           iterGCP != landmarks.end();
           ++iterGCP)  {
-          stream << iterGCP->second.X.transpose() << " 255 255 0" << '\n';
+          stream << iterGCP->second.X.transpose().cast<float>() << " 255 255 0" << '\n';
         }
       }
